+++ tests/testthat/test-geeglm.R
@@ -1,7 +1,4 @@
 if (require("testthat") && require("ggeffects") && require("geepack")) {
-
-  context("geeglm")
-
   data(dietox)
 
   m1 <- geeglm(
@@ -14,11 +11,11 @@ if (require("testthat") && require("ggeffects") && require("geepack")) {
 
   test_that("ggpredict", {
     p <- ggpredict(m1, c("Cu", "Time"))
-    expect_equal(p$predicted[1], 35.35779, tolerance = 1e-3)
+    expect_equal(p$predicted[1], 35.47711, tolerance = 1e-2)
   })
 
   test_that("ggemmeans", {
     p <- ggemmeans(m1, c("Cu", "Time"))
-    expect_equal(p$predicted[1], 35.35779, tolerance = 1e-3)
+    expect_equal(p$predicted[1], 35.47711, tolerance = 1e-2)
   })
 }